Corrected the enum minor error, added TTCard and TTDeck objects
The test cases came up correctly.  It should be possible to create a
card and a deck now.  With array, we can use the build-in push and pop
methods.  for instance, var array = [];
array.push(object);  thus we don't need to create the add card method.
although we might need to create the shuffle deck method later